deploy-service/src/proton-rds-sdk-py/tests/requirements.txt
allure-pytest
PyMySQL
pytest